Script for beta config; unique data dir, in-app env/type display
To release a beta build, run `npm prepare-beta` and then go through all
the standard release activities.

This adds a 'type' property to package.json to change the behavior of
the app, and changes the deploy targets. The key behavior change is a
new location for application data - for example, if in beta mode,
production data will be saved in 'Signal-beta' instead of just 'Signal.'

As part of this, three new bits of data are shown across the app:
  - Environment (development or test, not shown if production)
  - Type (only 'beta' for now; not shown if falsy)
  - App Instance (disabled in production; used for multiple accounts)

These are shown in:
  - The window title - all three
  - The about window - all three
  - The header added to the debug log - environment and type